Extract core module
