+++ rustdss_core/src/base_logic.rs
