+++ rustdss_core/src/lists.rs
